move provider and app file
