--- src/components/App.js
@@ -1,14 +1,0 @@
-import GlobalStatusBar from './GlobalStatusBar/GlobalStatusBar';
-import Main from './Main';
-import './App.css';
-
-const App = () => {
-  return (
-    <>
-      <GlobalStatusBar />
-      <Main />
-    </>
-  );
-};
-
-export default App;
