+++ src/index.js
@@ -1,18 +1,14 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import '@astrouxds/astro-web-components/dist/astro-web-components/astro-web-components.css';
-
 import reportWebVitals from './reportWebVitals';
-import AppProvider from './providers/AppProvider';
-import App from './components/App';
+import App from './App';
 import './index.css';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <React.StrictMode>
-    <AppProvider>
-      <App />
-    </AppProvider>
+    <App />
   </React.StrictMode>
 );
